astype fix 3

+++ mozaik/sheets/population_selector.py
@@ -92,7 +92,7 @@ class RCRandomPercentage(PopulationSelector):
         
     def generate_idd_list_of_neurons(self):
         if isinstance(self.sheet.pop.all_cells, list):
-            z = numpy.array(self.sheet.pop.all_cells, dtype=numpy.int32)
+            z = numpy.array(self.sheet.pop.all_cells, dtype=numpy.float32)
         else:
             z = self.sheet.pop.all_cells.astype(int)
         mozaik.rng.shuffle(z)
